chore(daemon): rename internal_proto -> server
The name was shadowing the shadow of the same name in `lib` and I
don’t even.
