+++ src/daemon/server.rs
